added a 'cache buster' to keep github actions from failing.
- Github Actions runs the main.py script to poll the EVE Online API
every 15 minutes.
- If the API does not return new results, the index.html file will not
be updated.
- If the index.html file is not updated, the resulting 'commit' action
in the Github Actions workflow will fail as there are no new changes to
commit.
- If the Github Actions Workflow fails, I will get an email stating as
such. Every 15 minutes.
- Although this ultimately has no impact on the accuracy of the website
content (it will -eventually- update, and be correct), receiving mail
informing me of a 'failure' of the workflow is confusing especially to
outsiders.
- Enter the 'Cache Buster'! Mine is a variant of a technique used by
frontend web devs to force clients to pull new versions of an asset from
a cache by dynamically changing said filename through the use of a
timestamp/hash of some kind. When client browsers detect the renamed
asset, it will discard the cached version in favor of a new one from the
origin.
- My cache buster takes the form of a timestamp via `int(time.time())`
inserted into index.html by way of a hidden input field at the bottom
of the page. This way, git detects a change and accepts the subsequent
commit.
- Since there is a valid commit on hand, the Workflow will succeed.
- When the Workflow succeeds I will receive a message confirming that